ExpertGPT/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
browser-cookie3==0.19.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
curl-cffi==0.5.7
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
g4f==0.1.4.0
idna==3.4
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
lz4==4.3.2
MarkupSafe==2.1.3
multidict==6.0.4
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
PyExecJS==1.5.1
pyjsparser==2.7.1
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzlocal==5.0.1
urllib3==2.0.5
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
